requirements.txt
pdfminer.six
pytesseract
pdf2image
tkinterdnd2
Pillow
